requirements.txt
cycler==0.12.1
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pandas==2.2.1
ply==3.11
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sympy==1.12
threadpoolctl==3.3.0
torch==2.2.1
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
webencodings==0.5.1
pyyaml==6.0.1